requirements.txt
requests
pandas
numpy
selenium
python-dotenv
openpyxl
beautifulsoup4
python-inquirer
inquirer